Download Google Trends data on Covid-19

Downloads Google Trends data (https://trends.google.com/trends/) about the 2020 search activity for a given search term at global and country levels. The search term defaults to "coronavirus" to reflect the relative public attention to the Covid-19 pandemic.

download_google_trends_data(
  search_term = "coronavirus",
  type = "country_day",
  silent = FALSE,
  cached = FALSE
)

Arguments

search_term
Defaults to "coronavirus".

type
The type of data that you want to retrieve. Can be any subset of

"country": Relative search activity at the global level, reporting by country.

"country_day": Relative search activity at the country level, reporting by country.

"region": Relative search activity at the country level, reporting by region.

"city": Relative search activity at the country level, reporting by city.

silent
Whether you want the function to send some status messages to the console. Might be informative as downloading will take some time and thus defaults to TRUE.

cached
Whether you want to download the cached version of the data from the tidycovid19 Github repository instead of retrieving the data from the authorative source. Downloading the cached version is faster and the cache is updated daily. Defaults to FALSE.

Value

If only one type was selected, a data frame containing the data. Otherwise, a list containing the desired data frames ordered as in type.

Details

Uses the gtrendsR package. Please note that Google Trends only reports relative search volume. For each data frame, the values are standardized so that the observations with the highest search volume gets a score of 100 and the other scores of the data frame are relative to that. This implies that comparisons across data frames are not feasible. When Google Trends reports a score of "<1" this is translated to 0.5 in the data.

Examples

df <- download_google_trends_data(type = "country", silent = TRUE, cached = TRUE)
df %>%
  dplyr::select(iso3c, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>    iso3c gtrends_score
#> 1    ITA           100
#> 2    ESP            89
#> 3    QAT            86
#> 4    FRA            72
#> 5    IRL            71
#> 6    GBR            68
#> 7    NPL            67
#> 8    ARE            67
#> 9    PER            65
#> 10   PAN            60
#> 11   ZAF            60
#> 12   CHE            58
#> 13   GHA            57
#> 14   AUS            56
#> 15   CAN            54
#> 16   USA            53
#> 17   NZL            53
#> 18   COL            53
#> 19   URY            52
#> 20   PRY            50
#> 21   ARG            50
#> 22   BOL            49
#> 23   SGP            46
#> 24   ROU            46
#> 25   CHL            45
#> 26   ECU            45
#> 27   GTM            43
#> 28   CRI            43
#> 29   AUT            41
#> 30   MEX            41
#> 31   DEU            41
#> 32   BEL            41
#> 33   KEN            41
#> 34   VEN            39
#> 35   IND            38
#> 36   DOM            38
#> 37   BGD            36
#> 38   PAK            36
#> 39   PRT            36
#> 40   MAR            34
#> 41   NGA            30
#> 42   BRA            30
#> 43   NLD            24
#> 44   DZA            23
#> 45   MYS            21
#> 46   SWE            19
#> 47   SAU            19
#> 48   DNK            19
#> 49   PHL            16
#> 50   NOR            13
#> 51   CZE             9
#> 52   GRC             9
#> 53   EGY             6
#> 54   POL             5
#> 55   THA             4
#> 56   IDN             4
#> 57   TUR             4
#> 58   VNM             2
#> 59   JPN             1

lst <- download_google_trends_data(type = c("region", "city"), silent = TRUE, cached = TRUE)
lst[[1]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(region, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>                    region gtrends_score
#> 1                Saarland           100
#> 2       Baden-Württemberg            96
#> 3               Thuringia            93
#> 4                  Berlin            93
#> 5           Saxony-Anhalt            92
#> 6  North Rhine-Westphalia            91
#> 7             Brandenburg            90
#> 8                 Bavaria            89
#> 9                  Bremen            89
#> 10   Rhineland-Palatinate            87
#> 11                 Saxony            87
#> 12     Schleswig-Holstein            87
#> 13                  Hesse            84
#> 14           Lower Saxony            82
#> 15                Hamburg            82
#> 16 Mecklenburg-Vorpommern            81

lst[[2]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(city, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>               city gtrends_score
#> 1           Bochum           100
#> 2          Krefeld            97
#> 3         Dortmund            90
#> 4         Duisburg            89
#> 5           Bremen            87
#> 6            Essen            87
#> 7  Mönchengladbach            87
#> 8           Berlin            87
#> 9         Mannheim            87
#> 10       Nuremberg            85
#> 11      Düsseldorf            85
#> 12         Dresden            84
#> 13          Munich            82
#> 14       Wuppertal            82
#> 15       Stuttgart            81
#> 16         Hamburg            80
#> 17            Bonn            79
#> 18         Leipzig            78
#> 19         Hanover            78
#> 20       Karlsruhe            78
#> 21       Wiesbaden            77
#> 22       Bielefeld            77
#> 23         Cologne            75
#> 24         Münster            74
#> 25       Frankfurt            70
#> 26     Drochtersen            65