Download Google Trends data on Covid-19

Downloads Google Trends data (https://trends.google.com/trends/) about the 2020 search activity for a given search term at global and country levels. The search term defaults to "coronavirus" to reflect the relative public attention to the Covid-19 pandemic.

download_google_trends_data(search_term = "coronavirus",
  type = "country_day", silent = FALSE, cached = FALSE)

Arguments

search_term
Defaults to "coronavirus".

type
The type of data that you want to retrieve. Can be any subset of

"country": Relative search activity at the global level, reporting by country.

"country_day": Relative search activity at the country level, reporting by country.

"region": Relative search activity at the country level, reporting by region.

"city": Relative search activity at the country level, reporting by city.

silent
Whether you want the function to send some status messages to the console. Might be informative as downloading will take some time and thus defaults to TRUE.

cached
Whether you want to download the cached version of the data from the tidycovid19 Github repository instead of retrieving the data from the authorative source. Downloading the cached version is faster and the cache is updated daily. Defaults to FALSE.

Value

If only one type was selected, a data frame containing the data. Otherwise, a list containing the desired data frames ordered as in type.

Details

Uses the gtrendsR package. Please note that Google Trends only reports relative search volume. For each data frame, the values are standardized so that the observations with the highest search volume gets a score of 100 and the other scores of the data frame are relative to that. This implies that comparisons across data frames are not feasible. When Google Trends reports a score of "<1" this is translated to 0.5 in the data.

Examples

df <- download_google_trends_data(type = "country", silent = TRUE, cached = TRUE)
df %>%
  dplyr::select(iso3c, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>    iso3c gtrends_score
#> 1    ITA           100
#> 2    ESP            79
#> 3    FRA            64
#> 4    IRL            61
#> 5    GBR            60
#> 6    QAT            55
#> 7    CHE            55
#> 8    PAN            48
#> 9    CAN            48
#> 10   ARE            47
#> 11   AUS            47
#> 12   PER            47
#> 13   USA            47
#> 14   NZL            46
#> 15   ZAF            45
#> 16   URY            43
#> 17   ROU            40
#> 18   ARG            40
#> 19   COL            39
#> 20   AUT            38
#> 21   DEU            38
#> 22   SGP            38
#> 23   PRT            37
#> 24   BOL            37
#> 25   CHL            36
#> 26   BEL            35
#> 27   CRI            35
#> 28   ECU            35
#> 29   VEN            30
#> 30   GTM            30
#> 31   MEX            29
#> 32   IND            28
#> 33   KEN            27
#> 34   DOM            27
#> 35   BRA            23
#> 36   MAR            23
#> 37   PAK            22
#> 38   NLD            22
#> 39   NGA            21
#> 40   BGD            21
#> 41   SWE            18
#> 42   MYS            17
#> 43   DNK            15
#> 44   DZA            15
#> 45   NOR            13
#> 46   PHL            12
#> 47   HKG            11
#> 48   SAU            11
#> 49   CZE            10
#> 50   GRC             7
#> 51   POL             4
#> 52   TUR             3
#> 53   THA             3
#> 54   IDN             3
#> 55   VNM             2
#> 56   JPN             1

lst <- download_google_trends_data(type = c("region", "city"), silent = TRUE, cached = TRUE)
lst[[1]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(region, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>                    region gtrends_score
#> 1                Saarland           100
#> 2       Baden-Württemberg            98
#> 3                  Berlin            96
#> 4                  Bremen            94
#> 5  North Rhine-Westphalia            92
#> 6                 Bavaria            92
#> 7           Saxony-Anhalt            89
#> 8    Rhineland-Palatinate            89
#> 9               Thuringia            89
#> 10                Hamburg            88
#> 11            Brandenburg            88
#> 12                  Hesse            87
#> 13     Schleswig-Holstein            86
#> 14                 Saxony            84
#> 15           Lower Saxony            83
#> 16 Mecklenburg-Vorpommern            78

lst[[2]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(city, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>           city gtrends_score
#> 1     Augsburg           100
#> 2       Bochum            98
#> 3     Mannheim            97
#> 4    Nuremberg            94
#> 5       Bremen            92
#> 6       Berlin            92
#> 7        Essen            92
#> 8     Duisburg            92
#> 9     Dortmund            91
#> 10   Wuppertal            91
#> 11   Stuttgart            90
#> 12      Munich            89
#> 13  Düsseldorf            89
#> 14   Wiesbaden            87
#> 15   Karlsruhe            87
#> 16     Hamburg            86
#> 17   Bielefeld            85
#> 18     Cologne            84
#> 19     Hanover            83
#> 20     Dresden            82
#> 21        Bonn            82
#> 22     Leipzig            82
#> 23     Münster            78
#> 24   Frankfurt            75
#> 25 Drochtersen            72