Download Google Trends data on Covid-19

Downloads Google Trends data (https://trends.google.com/trends/) about the 2020 search activity for a given search term at global and country levels. The search term defaults to "coronavirus" to reflect the relative public attention to the Covid-19 pandemic.

download_google_trends_data(search_term = "coronavirus",
  type = "country_day", silent = FALSE, cached = FALSE)

Arguments

search_term
Defaults to "coronavirus".

type
The type of data that you want to retrieve. Can be any subset of

"country": Relative search activity at the global level, reporting by country.

"country_day": Relative search activity at the country level, reporting by country.

"region": Relative search activity at the country level, reporting by region.

"city": Relative search activity at the country level, reporting by city.

silent
Whether you want the function to send some status messages to the console. Might be informative as downloading will take some time and thus defaults to TRUE.

cached
Whether you want to download the cached version of the data from the tidycovid19 Github repository instead of retrieving the data from the authorative source. Downloading the cached version is faster and the cache is updated daily. Defaults to FALSE.

Value

If only one type was selected, a data frame containing the data. Otherwise, a list containing the desired data frames ordered as in type.

Details

Uses the gtrendsR package. Please note that Google Trends only reports relative search volume. For each data frame, the values are standardized so that the observations with the highest search volume gets a score of 100 and the other scores of the data frame are relative to that. This implies that comparisons across data frames are not feasible. When Google Trends reports a score of "<1" this is translated to 0.5 in the data.

Examples

df <- download_google_trends_data(type = "country", silent = TRUE, cached = TRUE)
df %>%
  dplyr::select(iso3c, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>    iso3c gtrends_score
#> 1    ITA           100
#> 2    ESP            81
#> 3    FRA            67
#> 4    IRL            66
#> 5    GBR            64
#> 6    QAT            61
#> 7    CHE            56
#> 8    PER            54
#> 9    ARE            53
#> 10   PAN            51
#> 11   USA            50
#> 12   CAN            50
#> 13   AUS            49
#> 14   NZL            49
#> 15   ZAF            48
#> 16   URY            45
#> 17   COL            45
#> 18   PRY            44
#> 19   ARG            44
#> 20   ROU            43
#> 21   AUT            40
#> 22   BOL            39
#> 23   DEU            39
#> 24   CHL            39
#> 25   PRT            39
#> 26   SGP            39
#> 27   ECU            38
#> 28   CRI            38
#> 29   BEL            37
#> 30   GTM            34
#> 31   MEX            33
#> 32   VEN            33
#> 33   IND            30
#> 34   DOM            29
#> 35   KEN            29
#> 36   BRA            27
#> 37   MAR            27
#> 38   PAK            24
#> 39   NLD            24
#> 40   BGD            23
#> 41   NGA            23
#> 42   SWE            18
#> 43   DZA            18
#> 44   MYS            17
#> 45   DNK            17
#> 46   NOR            14
#> 47   SAU            13
#> 48   PHL            12
#> 49   HKG            12
#> 50   CZE             9
#> 51   GRC             8
#> 52   POL             4
#> 53   TUR             3
#> 54   THA             3
#> 55   IDN             3
#> 56   VNM             2
#> 57   JPN             1

lst <- download_google_trends_data(type = c("region", "city"), silent = TRUE, cached = TRUE)
lst[[1]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(region, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>                    region gtrends_score
#> 1                Saarland           100
#> 2       Baden-Württemberg            97
#> 3                 Bavaria            91
#> 4  North Rhine-Westphalia            90
#> 5                  Berlin            90
#> 6                  Bremen            88
#> 7               Thuringia            87
#> 8             Brandenburg            87
#> 9           Saxony-Anhalt            87
#> 10                Hamburg            86
#> 11   Rhineland-Palatinate            86
#> 12                 Saxony            85
#> 13     Schleswig-Holstein            84
#> 14                  Hesse            83
#> 15           Lower Saxony            82
#> 16 Mecklenburg-Vorpommern            78

lst[[2]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(city, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>               city gtrends_score
#> 1           Bochum           100
#> 2  Mönchengladbach            95
#> 3            Essen            95
#> 4           Berlin            95
#> 5         Dortmund            95
#> 6        Wuppertal            94
#> 7       Düsseldorf            93
#> 8         Mannheim            93
#> 9         Duisburg            93
#> 10       Stuttgart            92
#> 11          Munich            92
#> 12          Bremen            91
#> 13       Nuremberg            91
#> 14         Hamburg            88
#> 15         Dresden            87
#> 16       Karlsruhe            86
#> 17       Wiesbaden            86
#> 18         Cologne            84
#> 19         Leipzig            84
#> 20            Bonn            84
#> 21         Hanover            84
#> 22       Bielefeld            82
#> 23         Münster            79
#> 24       Frankfurt            77
#> 25     Drochtersen            73