Download Google Trends data on Covid-19

Downloads Google Trends data (https://trends.google.com/trends/) about the 2020 search activity for a given search term at global and country levels. The search term defaults to "coronavirus" to reflect the relative public attention to the Covid-19 pandemic.

download_google_trends_data(
  search_term = "coronavirus",
  type = "country_day",
  silent = FALSE,
  cached = FALSE
)

Arguments

search_term
Defaults to "coronavirus".

type
The type of data that you want to retrieve. Can be any subset of

"country": Relative search activity at the global level, reporting by country.

"country_day": Relative search activity at the country level, reporting by country.

"region": Relative search activity at the country level, reporting by region.

"city": Relative search activity at the country level, reporting by city.

silent
Whether you want the function to send some status messages to the console. Might be informative as downloading will take some time and thus defaults to TRUE.

cached
Whether you want to download the cached version of the data from the tidycovid19 Github repository instead of retrieving the data from the authorative source. Downloading the cached version is faster and the cache is updated daily. Defaults to FALSE.

Value

If only one type was selected, a data frame containing the data. Otherwise, a list containing the desired data frames ordered as in type.

Details

Uses the gtrendsR package. Please note that Google Trends only reports relative search volume. For each data frame, the values are standardized so that the observations with the highest search volume gets a score of 100 and the other scores of the data frame are relative to that. This implies that comparisons across data frames are not feasible. When Google Trends reports a score of "<1" this is translated to 0.5 in the data.

Examples

df <- download_google_trends_data(type = "country", silent = TRUE, cached = TRUE)
df %>%
  dplyr::select(iso3c, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>    iso3c gtrends_score
#> 1    ITA           100
#> 2    QAT            87
#> 3    ESP            87
#> 4    NPL            76
#> 5    FRA            71
#> 6    IRL            69
#> 7    GBR            69
#> 8    ARE            69
#> 9    PER            63
#> 10   ZAF            60
#> 11   GHA            57
#> 12   PAN            57
#> 13   CHE            56
#> 14   AUS            53
#> 15   USA            53
#> 16   COL            53
#> 17   CAN            52
#> 18   NZL            52
#> 19   URY            50
#> 20   BOL            50
#> 21   ARG            50
#> 22   PRY            50
#> 23   SGP            45
#> 24   ECU            45
#> 25   CHL            43
#> 26   ROU            43
#> 27   GTM            43
#> 28   CRI            43
#> 29   IND            40
#> 30   KEN            40
#> 31   DEU            40
#> 32   MEX            40
#> 33   AUT            39
#> 34   BEL            39
#> 35   VEN            39
#> 36   BGD            39
#> 37   DOM            38
#> 38   PAK            36
#> 39   PRT            35
#> 40   MAR            33
#> 41   NGA            30
#> 42   BRA            29
#> 43   DZA            23
#> 44   NLD            23
#> 45   MYS            21
#> 46   SAU            19
#> 47   SWE            18
#> 48   DNK            18
#> 49   PHL            16
#> 50   NOR            14
#> 51   CZE             9
#> 52   GRC             8
#> 53   EGY             7
#> 54   POL             4
#> 55   THA             4
#> 56   IDN             4
#> 57   TUR             4
#> 58   VNM             2
#> 59   JPN             1

lst <- download_google_trends_data(type = c("region", "city"), silent = TRUE, cached = TRUE)
lst[[1]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(region, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>                    region gtrends_score
#> 1                Saarland           100
#> 2           Saxony-Anhalt            96
#> 3       Baden-Württemberg            94
#> 4               Thuringia            93
#> 5  North Rhine-Westphalia            92
#> 6                  Berlin            92
#> 7                  Bremen            92
#> 8                 Bavaria            91
#> 9             Brandenburg            91
#> 10   Rhineland-Palatinate            89
#> 11                 Saxony            88
#> 12     Schleswig-Holstein            88
#> 13 Mecklenburg-Vorpommern            86
#> 14           Lower Saxony            84
#> 15                Hamburg            82
#> 16                  Hesse            82

lst[[2]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(city, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>           city gtrends_score
#> 1      Krefeld           100
#> 2       Bochum            96
#> 3     Augsburg            93
#> 4        Essen            92
#> 5     Duisburg            91
#> 6     Dortmund            90
#> 7       Bremen            90
#> 8       Berlin            88
#> 9    Nuremberg            88
#> 10    Chemnitz            88
#> 11   Wuppertal            88
#> 12    Mannheim            88
#> 13      Munich            84
#> 14  Düsseldorf            83
#> 15   Stuttgart            83
#> 16   Wiesbaden            82
#> 17     Hamburg            80
#> 18   Karlsruhe            80
#> 19     Dresden            80
#> 20     Leipzig            78
#> 21     Cologne            78
#> 22        Bonn            77
#> 23     Hanover            76
#> 24     Münster            73
#> 25   Frankfurt            70
#> 26 Drochtersen            68