Download Google Trends data on Covid-19

Downloads Google Trends data (https://trends.google.com/trends/) about the 2020 search activity for a given search term at global and country levels. The search term defaults to "coronavirus" to reflect the relative public attention to the Covid-19 pandemic.

download_google_trends_data(search_term = "coronavirus",
  type = "country_day", silent = FALSE, cached = FALSE)

Arguments

search_term
Defaults to "coronavirus".

type
The type of data that you want to retrieve. Can be any subset of

"country": Relative search activity at the global level, reporting by country.

"country_day": Relative search activity at the country level, reporting by country.

"region": Relative search activity at the country level, reporting by region.

"city": Relative search activity at the country level, reporting by city.

silent
Whether you want the function to send some status messages to the console. Might be informative as downloading will take some time and thus defaults to TRUE.

cached
Whether you want to download the cached version of the data from the tidycovid19 Github repository instead of retrieving the data from the authorative source. Downloading the cached version is faster and the cache is updated daily. Defaults to FALSE.

Value

If only one type was selected, a data frame containing the data. Otherwise, a list containing the desired data frames ordered as in type.

Details

Uses the gtrendsR package. Please note that Google Trends only reports relative search volume. For each data frame, the values are standardized so that the observations with the highest search volume gets a score of 100 and the other scores of the data frame are relative to that. This implies that comparisons across data frames are not feasible. When Google Trends reports a score of "<1" this is translated to 0.5 in the data.

Examples

df <- download_google_trends_data(type = "country", silent = TRUE, cached = TRUE)
df %>%
  dplyr::select(iso3c, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>    iso3c gtrends_score
#> 1    ITA           100
#> 2    ESP            76
#> 3    FRA            63
#> 4    IRL            59
#> 5    GBR            56
#> 6    CHE            53
#> 7    QAT            50
#> 8    PAN            48
#> 9    AUS            46
#> 10   NZL            45
#> 11   USA            45
#> 12   CAN            45
#> 13   PER            44
#> 14   ARE            43
#> 15   ZAF            41
#> 16   URY            40
#> 17   ROU            39
#> 18   ARG            38
#> 19   AUT            38
#> 20   PRT            38
#> 21   SGP            37
#> 22   COL            37
#> 23   DEU            37
#> 24   CHL            35
#> 25   ECU            34
#> 26   BOL            34
#> 27   BEL            33
#> 28   CRI            33
#> 29   VEN            30
#> 30   GTM            29
#> 31   MEX            26
#> 32   IND            24
#> 33   KEN            24
#> 34   DOM            23
#> 35   BRA            23
#> 36   NLD            22
#> 37   NGA            21
#> 38   PAK            21
#> 39   MAR            20
#> 40   SWE            17
#> 41   BGD            17
#> 42   MYS            15
#> 43   DNK            15
#> 44   NOR            12
#> 45   PHL            12
#> 46   CZE            10
#> 47   SAU             9
#> 48   GRC             7
#> 49   POL             4
#> 50   TUR             3
#> 51   THA             3
#> 52   IDN             2
#> 53   JPN             1

lst <- download_google_trends_data(type = c("region", "city"), silent = TRUE, cached = TRUE)
lst[[1]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(region, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>                    region gtrends_score
#> 1                Saarland           100
#> 2       Baden-Württemberg            97
#> 3                  Berlin            97
#> 4                 Bavaria            91
#> 5                  Bremen            90
#> 6  North Rhine-Westphalia            89
#> 7    Rhineland-Palatinate            88
#> 8                 Hamburg            86
#> 9               Thuringia            85
#> 10     Schleswig-Holstein            85
#> 11          Saxony-Anhalt            84
#> 12                  Hesse            84
#> 13            Brandenburg            84
#> 14                 Saxony            84
#> 15           Lower Saxony            82
#> 16 Mecklenburg-Vorpommern            73

lst[[2]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(city, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>           city gtrends_score
#> 1      Krefeld           100
#> 2       Bochum            98
#> 3     Mannheim            93
#> 4       Berlin            92
#> 5        Essen            92
#> 6     Duisburg            92
#> 7     Dortmund            90
#> 8       Munich            89
#> 9    Nuremberg            89
#> 10   Wuppertal            88
#> 11   Stuttgart            87
#> 12  Düsseldorf            87
#> 13   Karlsruhe            87
#> 14      Bremen            85
#> 15     Hamburg            84
#> 16     Dresden            82
#> 17   Wiesbaden            82
#> 18     Hanover            81
#> 19     Cologne            80
#> 20        Bonn            78
#> 21     Leipzig            77
#> 22     Münster            75
#> 23   Frankfurt            73
#> 24 Drochtersen            70