Download Google Trends data on Covid-19

Downloads Google Trends data (https://trends.google.com/trends/) about the 2020 search activity for a given search term at global and country levels. The search term defaults to "coronavirus" to reflect the relative public attention to the Covid-19 pandemic.

download_google_trends_data(search_term = "coronavirus",
  type = "country_day", silent = FALSE, cached = FALSE)

Arguments

search_term
Defaults to "coronavirus".

type
The type of data that you want to retrieve. Can be any subset of

"country": Relative search activity at the global level, reporting by country.

"country_day": Relative search activity at the country level, reporting by country.

"region": Relative search activity at the country level, reporting by region.

"city": Relative search activity at the country level, reporting by city.

silent
Whether you want the function to send some status messages to the console. Might be informative as downloading will take some time and thus defaults to TRUE.

cached
Whether you want to download the cached version of the data from the tidycovid19 Github repository instead of retrieving the data from the authorative source. Downloading the cached version is faster and the cache is updated daily. Defaults to FALSE.

Value

If only one type was selected, a data frame containing the data. Otherwise, a list containing the desired data frames ordered as in type.

Details

Uses the gtrendsR package. Please note that Google Trends only reports relative search volume. For each data frame, the values are standardized so that the observations with the highest search volume gets a score of 100 and the other scores of the data frame are relative to that. This implies that comparisons across data frames are not feasible. When Google Trends reports a score of "<1" this is translated to 0.5 in the data.

Examples

df <- download_google_trends_data(type = "country", silent = TRUE, cached = TRUE)
df %>%
  dplyr::select(iso3c, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>    iso3c gtrends_score
#> 1    ITA           100
#> 2    ESP            77
#> 3    FRA            62
#> 4    IRL            59
#> 5    CHE            57
#> 6    GBR            55
#> 7    PAN            46
#> 8    PER            44
#> 9    USA            43
#> 10   CAN            43
#> 11   AUS            42
#> 12   NZL            42
#> 13   QAT            41
#> 14   AUT            40
#> 15   URY            39
#> 16   PRT            39
#> 17   DEU            39
#> 18   ROU            38
#> 19   SGP            36
#> 20   ARG            36
#> 21   ZAF            36
#> 22   ARE            35
#> 23   COL            34
#> 24   BEL            33
#> 25   CHL            33
#> 26   CRI            32
#> 27   BOL            32
#> 28   ECU            32
#> 29   GTM            27
#> 30   VEN            27
#> 31   MEX            25
#> 32   NLD            23
#> 33   DOM            22
#> 34   BRA            22
#> 35   KEN            21
#> 36   IND            19
#> 37   SWE            18
#> 38   MAR            18
#> 39   NGA            17
#> 40   PAK            16
#> 41   DNK            16
#> 42   MYS            15
#> 43   NOR            14
#> 44   PHL            12
#> 45   BGD            12
#> 46   HKG            11
#> 47   CZE            10
#> 48   GRC             7
#> 49   SAU             7
#> 50   POL             4
#> 51   TUR             3
#> 52   THA             3
#> 53   IDN             2
#> 54   JPN             1

lst <- download_google_trends_data(type = c("region", "city"), silent = TRUE, cached = TRUE)
lst[[1]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(region, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>                    region gtrends_score
#> 1                Saarland           100
#> 2       Baden-Württemberg            99
#> 3                  Berlin            96
#> 4                 Bavaria            95
#> 5  North Rhine-Westphalia            91
#> 6                  Bremen            89
#> 7                 Hamburg            89
#> 8    Rhineland-Palatinate            88
#> 9             Brandenburg            87
#> 10          Saxony-Anhalt            86
#> 11                  Hesse            86
#> 12     Schleswig-Holstein            85
#> 13              Thuringia            85
#> 14                 Saxony            83
#> 15           Lower Saxony            82
#> 16 Mecklenburg-Vorpommern            76

lst[[2]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(city, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>               city gtrends_score
#> 1          Krefeld           100
#> 2        Stuttgart            96
#> 3         Augsburg            94
#> 4           Munich            93
#> 5           Bochum            92
#> 6         Mannheim            92
#> 7           Aachen            92
#> 8        Nuremberg            92
#> 9           Berlin            91
#> 10 Mönchengladbach            91
#> 11           Essen            91
#> 12        Freiburg            90
#> 13       Wuppertal            90
#> 14        Duisburg            89
#> 15        Dortmund            87
#> 16       Karlsruhe            86
#> 17          Bremen            86
#> 18      Düsseldorf            86
#> 19         Hamburg            85
#> 20       Wiesbaden            83
#> 21         Cologne            83
#> 22       Bielefeld            82
#> 23         Dresden            82
#> 24         Leipzig            80
#> 25         Hanover            80
#> 26            Bonn            79
#> 27         Münster            75
#> 28     Drochtersen            73
#> 29       Frankfurt            73