Download Google Trends data on Covid-19

Downloads Google Trends data (https://trends.google.com/trends/) about the 2020 search activity for a given search term at global and country levels. The search term defaults to "coronavirus" to reflect the relative public attention to the Covid-19 pandemic.

download_google_trends_data(
  search_term = "coronavirus",
  type = "country_day",
  countries = NULL,
  low_search_volume = FALSE,
  pause = TRUE,
  silent = FALSE,
  cached = FALSE
)

Arguments

search_term
Defaults to "coronavirus".

type
The type of data that you want to retrieve. Can be any subset of

"country": Relative search activity at the global level, reporting by country.

"country_day": Relative search activity at the country level, reporting by country.

"region": Relative search activity at the country level, reporting by region.

"city": Relative search activity at the country level, reporting by city.

Defaults to 'country_day'.

countries
A character vector of ISO3c country codes that you want to pull detailed data for. By default (countries = NULL) the code pulls detailed data for all countries the the global Google Trend request returns.

low_search_volume
Whether you want the include countries with low search volume. This increases the list of countries that are pulled considerably and also the risk that you hit a Google Trend search limit. Use with care. Defaults to FALSE.

pause
Whether you want the code to pause for a 2 to 5 seconds period between Google Trend API calls. As Google Trends has an unknown rate limit, this is probably a good idea and thus defaults to TRUE.

silent
Whether you want the function to send some status messages to the console. Might be informative as downloading will take some time and thus defaults to TRUE.

cached
Whether you want to download the cached version of the data from the tidycovid19 Github repository instead of retrieving the data from the authorative source. Downloading the cached version is faster and the cache is updated daily. Defaults to FALSE.

Value

If only one type was selected, a data frame containing the data. Otherwise, a list containing the desired data frames ordered as in type.

Details

Uses the gtrendsR package. Please note that Google Trends only reports relative search volume. For each data frame, the values are standardized so that the observations with the highest search volume gets a score of 100 and the other scores of the data frame are relative to that. This implies that comparisons across data frames are not feasible. When Google Trends reports a score of "<1" this is translated to 0.5 in the data.

Examples

df <- download_google_trends_data(type = "country", silent = TRUE, cached = TRUE)
df %>%
  dplyr::select(iso3c, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>    iso3c gtrends_score
#> 1    ITA           100
#> 2    ESP            82
#> 3    GBR            77
#> 4    QAT            76
#> 5    IRL            71
#> 6    FRA            66
#> 7    CHE            59
#> 8    URY            56
#> 9    ARE            55
#> 10   AUS            53
#> 11   ARG            52
#> 12   NPL            51
#> 13   CAN            46
#> 14   ZAF            46
#> 15   USA            44
#> 16   ROU            44
#> 17   COL            43
#> 18   PER            43
#> 19   NZL            42
#> 20   AUT            42
#> 21   PAN            40
#> 22   SGP            40
#> 23   BEL            37
#> 24   DEU            37
#> 25   CHL            37
#> 26   IND            34
#> 27   BOL            34
#> 28   CRI            32
#> 29   ECU            30
#> 30   MEX            30
#> 31   GTM            29
#> 32   VEN            29
#> 33   PRT            28
#> 34   MAR            27
#> 35   PAK            27
#> 36   LKA            26
#> 37   BGD            26
#> 38   BRA            25
#> 39   KEN            24
#> 40   DOM            24
#> 41   NLD            21
#> 42   DZA            20
#> 43   MYS            16
#> 44   NGA            16
#> 45   DNK            15
#> 46   SAU            15
#> 47   SWE            14
#> 48   ISR            11
#> 49   PHL            10
#> 50   GRC            10
#> 51   EGY             6
#> 52   POL             4
#> 53   THA             4
#> 54   TUR             3
#> 55   IDN             3
#> 56   RUS             2
#> 57   JPN             1

lst <- download_google_trends_data(type = c("region", "city"), silent = TRUE, cached = TRUE)
lst[[1]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(region, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>                    region gtrends_score
#> 1       Baden-Württemberg           100
#> 2               Thuringia           100
#> 3                Saarland            95
#> 4           Saxony-Anhalt            92
#> 5                 Bavaria            91
#> 6  North Rhine-Westphalia            90
#> 7             Brandenburg            89
#> 8  Mecklenburg-Vorpommern            88
#> 9                  Saxony            88
#> 10     Schleswig-Holstein            87
#> 11                 Bremen            86
#> 12                 Berlin            85
#> 13   Rhineland-Palatinate            84
#> 14           Lower Saxony            82
#> 15                  Hesse            78
#> 16                Hamburg            74

lst[[2]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(city, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>                  city gtrends_score
#> 1      Markt Schwaben           100
#> 2 Uhldingen-Mühlhofen            90
#> 3           Sauerlach            88
#> 4               Hagen            81
#> 5              Bochum            79
#> 6            Sulzbach            79
#> 7             Cottbus            75
#> 8           Merseburg            56