Download Google Trends data on Covid-19

Downloads Google Trends data (https://trends.google.com/trends/) about the 2020 search activity for a given search term at global and country levels. The search term defaults to "coronavirus" to reflect the relative public attention to the Covid-19 pandemic.

download_google_trends_data(
  search_term = "coronavirus",
  type = "country_day",
  countries = NULL,
  low_search_volume = FALSE,
  pause = TRUE,
  silent = FALSE,
  cached = FALSE
)

Arguments

search_term
Defaults to "coronavirus".

type
The type of data that you want to retrieve. Can be any subset of

"country": Relative search activity at the global level, reporting by country.

"country_day": Relative search activity at the country level, reporting by country.

"region": Relative search activity at the country level, reporting by region.

"city": Relative search activity at the country level, reporting by city.

Defaults to 'country_day'.

countries
A character vector of ISO3c country codes that you want to pull detailed data for. By default (countries = NULL) the code pulls detailed data for all countries the the global Google Trend request returns.

low_search_volume
Whether you want the include countries with low search volume. This increases the list of countries that are pulled considerably and also the risk that you hit a Google Trend search limit. Use with care. Defaults to FALSE.

pause
Whether you want the code to pause for a 2 to 5 seconds period between Google Trend API calls. As Google Trends has an unknown rate limit, this is probably a good idea and thus defaults to TRUE.

silent
Whether you want the function to send some status messages to the console. Might be informative as downloading will take some time and thus defaults to TRUE.

cached
Whether you want to download the cached version of the data from the tidycovid19 Github repository instead of retrieving the data from the authorative source. Downloading the cached version is faster and the cache is updated daily. Defaults to FALSE.

Value

If only one type was selected, a data frame containing the data. Otherwise, a list containing the desired data frames ordered as in type.

Details

Uses the gtrendsR package. Please note that Google Trends only reports relative search volume. For each data frame, the values are standardized so that the observations with the highest search volume gets a score of 100 and the other scores of the data frame are relative to that. This implies that comparisons across data frames are not feasible. When Google Trends reports a score of "<1" this is translated to 0.5 in the data.

Examples

df <- download_google_trends_data(type = "country", silent = TRUE, cached = TRUE)
df %>%
  dplyr::select(iso3c, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>    iso3c gtrends_score
#> 1    ITA           100
#> 2    ESP            85
#> 3    GBR            77
#> 4    QAT            76
#> 5    IRL            69
#> 6    FRA            61
#> 7    CHE            59
#> 8    URY            56
#> 9    ARE            52
#> 10   AUS            51
#> 11   NPL            49
#> 12   ARG            49
#> 13   ZAF            47
#> 14   PER            46
#> 15   CAN            46
#> 16   NZL            43
#> 17   COL            42
#> 18   ROU            41
#> 19   AUT            40
#> 20   USA            40
#> 21   PAN            40
#> 22   BEL            39
#> 23   SGP            39
#> 24   PRY            37
#> 25   DEU            37
#> 26   CHL            37
#> 27   IND            35
#> 28   BOL            34
#> 29   GHA            32
#> 30   CRI            32
#> 31   ECU            30
#> 32   MEX            30
#> 33   VEN            28
#> 34   GTM            27
#> 35   PRT            27
#> 36   BGD            27
#> 37   MAR            25
#> 38   PAK            25
#> 39   KEN            25
#> 40   LKA            24
#> 41   DOM            24
#> 42   BRA            23
#> 43   DZA            20
#> 44   NLD            18
#> 45   MYS            16
#> 46   NGA            15
#> 47   DNK            15
#> 48   SAU            15
#> 49   SWE            13
#> 50   ISR            10
#> 51   PHL            10
#> 52   GRC             9
#> 53   EGY             6
#> 54   POL             4
#> 55   THA             3
#> 56   TUR             3
#> 57   IDN             3
#> 58   RUS             2
#> 59   JPN             1

lst <- download_google_trends_data(type = c("region", "city"), silent = TRUE, cached = TRUE)
lst[[1]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(region, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>                    region gtrends_score
#> 1               Thuringia           100
#> 2                  Bremen            94
#> 3       Baden-Württemberg            94
#> 4                Saarland            93
#> 5           Saxony-Anhalt            92
#> 6  North Rhine-Westphalia            91
#> 7            Lower Saxony            88
#> 8                  Saxony            87
#> 9                 Bavaria            86
#> 10     Schleswig-Holstein            86
#> 11                 Berlin            85
#> 12   Rhineland-Palatinate            83
#> 13            Brandenburg            83
#> 14                Hamburg            80
#> 15                  Hesse            79
#> 16 Mecklenburg-Vorpommern            79

lst[[2]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(city, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>         city gtrends_score
#> 1  Wolfsburg           100
#> 2   Neustadt            89
#> 3 Angermünde            89
#> 4   Meschede            80