Download Google Trends data on Covid-19

Downloads Google Trends data (https://trends.google.com/trends/) about the 2020 search activity for a given search term at global and country levels. The search term defaults to "coronavirus" to reflect the relative public attention to the Covid-19 pandemic.

download_google_trends_data(
  search_term = "coronavirus",
  type = "country_day",
  countries = NULL,
  low_search_volume = FALSE,
  pause = TRUE,
  silent = FALSE,
  cached = FALSE
)

Arguments

search_term
Defaults to "coronavirus".

type
The type of data that you want to retrieve. Can be any subset of

"country": Relative search activity at the global level, reporting by country.

"country_day": Relative search activity at the country level, reporting by country.

"region": Relative search activity at the country level, reporting by region.

"city": Relative search activity at the country level, reporting by city.

Defaults to 'country_day'.

countries
A character vector of ISO3c country codes that you want to pull detailed data for. By default (countries = NULL) the code pulls detailed data for all countries the the global Google Trend request returns.

low_search_volume
Whether you want the include countries with low search volume. This increases the list of countries that are pulled considerably and also the risk that you hit a Google Trend search limit. Use with care. Defaults to FALSE.

pause
Whether you want the code to pause for a 2 to 5 seconds period between Google Trend API calls. As Google Trends has an unknown rate limit, this is probably a good idea and thus defaults to TRUE.

silent
Whether you want the function to send some status messages to the console. Might be informative as downloading will take some time and thus defaults to TRUE.

cached
Whether you want to download the cached version of the data from the tidycovid19 Github repository instead of retrieving the data from the authorative source. Downloading the cached version is faster and the cache is updated daily. Defaults to FALSE.

Value

If only one type was selected, a data frame containing the data. Otherwise, a list containing the desired data frames ordered as in type.

Details

Uses the gtrendsR package. Please note that Google Trends only reports relative search volume. For each data frame, the values are standardized so that the observations with the highest search volume gets a score of 100 and the other scores of the data frame are relative to that. This implies that comparisons across data frames are not feasible. When Google Trends reports a score of "<1" this is translated to 0.5 in the data.

Examples

df <- download_google_trends_data(type = "country", silent = TRUE, cached = TRUE)
df %>%
  dplyr::select(iso3c, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>    iso3c gtrends_score
#> 1    ITA           100
#> 2    ESP            87
#> 3    GBR            79
#> 4    QAT            78
#> 5    IRL            70
#> 6    FRA            67
#> 7    URY            60
#> 8    CHE            57
#> 9    ARE            56
#> 10   ARG            53
#> 11   AUS            50
#> 12   NPL            48
#> 13   PER            46
#> 14   CAN            45
#> 15   NZL            45
#> 16   ROU            43
#> 17   ZAF            43
#> 18   KWT            42
#> 19   PAN            42
#> 20   USA            42
#> 21   SGP            41
#> 22   AUT            41
#> 23   COL            40
#> 24   BEL            37
#> 25   CHL            37
#> 26   DEU            36
#> 27   PRY            35
#> 28   IND            33
#> 29   BOL            32
#> 30   CRI            32
#> 31   ECU            32
#> 32   VEN            31
#> 33   GTM            30
#> 34   MEX            30
#> 35   PRT            27
#> 36   BGD            27
#> 37   MAR            27
#> 38   PAK            27
#> 39   LKA            26
#> 40   BRA            25
#> 41   DOM            25
#> 42   KEN            24
#> 43   DZA            21
#> 44   NLD            19
#> 45   MYS            16
#> 46   SAU            16
#> 47   DNK            16
#> 48   NGA            15
#> 49   SWE            13
#> 50   PHL            11
#> 51   ISR            11
#> 52   GRC             9
#> 53   EGY             5
#> 54   POL             4
#> 55   THA             3
#> 56   TUR             3
#> 57   IDN             3
#> 58   VNM             2
#> 59   RUS             2
#> 60   JPN             1

lst <- download_google_trends_data(type = c("region", "city"), silent = TRUE, cached = TRUE)
lst[[1]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(region, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>                    region gtrends_score
#> 1               Thuringia           100
#> 2                Saarland           100
#> 3       Baden-Württemberg            98
#> 4            Lower Saxony            95
#> 5  North Rhine-Westphalia            94
#> 6                 Bavaria            94
#> 7             Brandenburg            94
#> 8                  Bremen            92
#> 9           Saxony-Anhalt            91
#> 10                Hamburg            89
#> 11                 Berlin            89
#> 12                 Saxony            89
#> 13   Rhineland-Palatinate            88
#> 14                  Hesse            86
#> 15 Mecklenburg-Vorpommern            85
#> 16     Schleswig-Holstein            84

lst[[2]] %>%
  dplyr::filter(iso3c == "DEU") %>%
  dplyr::select(city, gtrends_score) %>%
  dplyr::arrange(-gtrends_score)

#>          city gtrends_score
#> 1 Großostheim           100
#> 2    Dortmund            87
#> 3      Munich            80
#> 4     Netphen            76
#> 5 Bremerhaven            74
#> 6    Würselen            69
#> 7   Falkensee            63
#> 8     Giessen            58