shiny_covid19_spread {tidycovid19}	R Documentation

Explore the spread of Covid-19 interactively

Description

A shiny based web app that allows users to customize the plot_covid19_spread display. The display has been inpired by the displays created by John Burn-Murdoch for the Financial Times. You can use it to customize your plot_covid19_spread() display as it allows copying the plot generating code to the clipboard, thanks to the fine rclipboard package.

Usage

shiny_covid19_spread(plot_options = list(), ...)

Arguments

plot_options
A named list containing parameters to initialize the plot in the shiny app. All parameters of plot_covid19_spread() are supported.

...
Parameters that will be parsed on to runApp.

Details

Uses data from the Johns Hopkins University CSSE team on the spread of the SARS-CoV-2 virus and the Covid-19 pandemic (https://github.com/CSSEGISandData/COVID-19), from the ACAPS governmental measures database (https://www.acaps.org/covid19-government-measures-dataset) and from the World Bank (https://data.worldbank.org). See the documentation of the separate download functions of the package for more detail.

Examples

## Not run: 
  shiny_covid19_spread()

  df <- download_merged_data(cached = TRUE)
  shiny_covid19_spread(
    plot_options = list(data = df, highlight = c("USA", "ESP", "ITA"))
  )

## End(Not run)

[Package tidycovid19 version 0.0.0.9000 Index]