tidycovid19_data_sources {tidycovid19}	R Documentation

Documentation of tidycovid19 Data Sources

Description

Brief info on each data source included in tidycovid19

Usage

data(tidycovid19_data_sources)

Format

An object of class "data.frame".

Source

Information has been collected by data providers (URLs included in data frame). Descriptions of data cleaning steps provided by the author of this package.

Examples

data(tidycovid19_data_sources)
print(tidycovid19_data_sources)

[Package tidycovid19 version 0.0.0.9000 Index]