Documentation of tidycovid19 Data Sources

Brief info on each data source included in tidycovid19

data(tidycovid19_data_sources)

Format

An object of class "data.frame".

Source

Information has been collected by data providers (URLs included in data frame). Descriptions of data cleaning steps provided by the author of this package.

Examples

data(tidycovid19_data_sources)
print(tidycovid19_data_sources)

#>               id                    function_name
#> 1       jhu_csse download_jhu_csse_covid19_data()
#> 2   ecdc_covid19     download_ecdc_covid19_data()
#> 3      owid_data             download_owid_data()
#> 4          wbank            download_wbank_data()
#> 5      acaps_npi        download_acaps_npi_data()
#> 6     oxford_npi       download_oxford_npi_data()
#> 7      apple_mtr        download_apple_mtr_data()
#> 8     google_cmr       download_google_cmr_data()
#> 9  google_trends    download_google_trends_data()
#> 10        merged           download_merged_data()
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          description
#> 1                                                                                                                                                                                                                                                                                                                                                                                                               The COVID-19 Data Repository by the Johns Hopkins University Center for Systems Science and Engineering (JHU CSSE) relies upon publicly available data from multiple sources that do not always agree. It is updated daily. The data comes in three data frames that you can select by the 'type' parameter. The 'country' data frame contains the global country-level data reported by JHU CSSE by aggregating over the regional data for countries that have regional data available. The 'country_region' data frame provides regional data for the countries that have regional data available (mosty Australia, Canada and China). The 'us_county' data frame reports the data for the U.S. at the county level. Please note: JHU stopped updating the data on March 10, 2023.
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      Country-level weekly data on new cases and deaths provided by the European Centre for Disease Prevention and Control (ECDC). The data was updated daily until 2020-12-14 and contains the latest available public data on the number of new Covid-19 cases reported per week and per country.
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The Our World in Data team systematically collects data on Covid-19 testing, hospitalizations, and vaccinations from multiple national sources. Data points are collected with varying frequency across countries. The definition on what consitutes a 'test' varies, reflected by the variable 'tests_units' in the data frame. The vaccination data is currently only available based on ad hoc disclosures by a small set of countries.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 The data frame reports current country-level statistics from the World Bank. The regional and income level classifications are also provided by the World Bank. 'life_expectancy' is measured in years at birth and 'gdp_capita' is measured in 2010 US-$. The original World Bank data items are (in the order how they are represented in the data frame) 'SP.POP.TOTL', 'AG.LND.TOTL.K2', 'EN.POP.DNST', 'EN.URB.LCTY', 'SP.DYN.LE00.IN', 'NY.GDP.PCAP.KD'. When you set the parameter 'var_def' tot 'TRUE'. the data comes in a list containing two data frames. The first contains the actual data, the second contains variable definitions.
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             The #COVID19 Government Measures Dataset is provided by ACAPS. It puts together measures implemented by governments worldwide in response to the Coronavirus pandemic. Data collection includes secondary data review. The data is reported in event structure with an event reflecting a government measure. Measures are characterized as being either imposing/extending measures or lifting them and categorized in five categories with each category being split up in further sub-categories. Please note: ACAPS stopped updating the data on December 10, 2020
#> 6                                                                                                                                                                                                                                                                         The data on the Oxford Coronavirus Government Response Tracker (OxCGRT) on non-pharmaceutical interventions comes in two data frames that you can select by setting the 'type' parameter. The 'measures' data frame reports data on governmental response measures as reported by the Oxford OxCGRT team. It is tidied by arranging its content by measure. All original country-day observations that are either initial or represent a value (not note) change from the previous day are included. Economic measures (E1-E4) are not included. The 'index' data frame reports the 'Stringency Index' and the 'Legacy Stringency Index' as calculated by the OxCGRT team based on their governance response measures in a country-day structure. Please note: As indicated on the homepage of the project, too a large extend the data is no longer updated after December 31, 2022 while data review processes continue.
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Apple's Mobility Trend Reports reflect requests for directions in Apple Maps. The data frame is organized by country-day and its data are expressed as percentages relative to a baseline volume on January 13th, 2020. The data comes in three data frames that you can select by the 'type' parameter. The 'country' data frame contains country-day level data. The 'country_region' data frame provides regional data for regions for which Apple reports regional data. The 'country_city' data frame reports city-level data for cities for which Apple reports this data. Please note: Apple stopped providing this data on April 14, 2022
#> 8                                                                                                                                                                                                                                                                                                                                       Google's Community Mobility Reports chart movement trends over time across different categories of places such as retail and recreation, groceries and pharmacies, parks, transit stations, workplaces, and residential. They show how visits and length of stays at different places change in percentages compared to a baseline (the median value, for the corresponding day of the week, during the 5-week period Jan 3 to Feb 6, 2020). The data comes in three data frames that you can select by the 'type' parameter. The 'country' data frame contains country-day level data. The 'country_region' data frame provides regional data for the countries for which Google reports regional data. The 'us_county' data frame reports daily data for the U.S. at the county level. Please note: Google stopped providing that data on October 15, 2022
#> 9                                                                                                                                                                                                             Data are Google Search Volume (GSV) measures as provided by Google Trends API, with the default search term 'coronavirus'. The data comes in four data frames that you can select by the 'type' parameter and the sample period comprises Jan 1, 2020 up to date. The 'country' data frame lists GSV by country, to assess which country on average uses the search term most often over the sample period. The 'country-day' data frame reports daily search volume data for all countries that show up in the 'country' data frame. Each value is relative within country, meaning that values across countries cannot be compared directly. The 'region' and 'city' data frames list the relative GSV across regions and city within countries when provided by Google Trends. Keep in mind that within each data frame GSV are relative measures with a maximum of 100 indicating the highest search volume. This implies that GSV measures are not comparable across data frames.
#> 10 This data frame contains Covid-19 related data from multiple sources in a country-day structure. Data sources are JHU CSSE data on confirmed cases, deaths and recoveries (https://github.com/CSSEGISandData/COVID-19), 'Our World in Data' data (https://github.com/owid/covid-19-data/tree/master/public/data), ACAPS data on governmental measures (https://www.acaps.org/covid19-government-measures-dataset), Oxford Covid-19 Government Respoonse Tracker (https://github.com/OxCGRT/covid-policy-tracker), Apple's Mobility Trend Reports on Apple Map usage (https://www.apple.com/covid19/mobility), Google's Community Mobility Reports on individual movement trends (https://www.google.com/covid19/mobility/), Google Trends data on relative Google search volumes for the term 'coronavirus' (https://trends.google.com/) and country-level World Bank data on population (density), life expectancy and national income (https://data.worldbank.org). The data frame 'tidycovid19_variable_definitions' holds definitions for each variable in this data frame. The data frame 'tidycovid19_data_sources' contains more information on the data sources included in this package.
#>                                                                                            url
#> 1                                                   https://github.com/CSSEGISandData/COVID-19
#> 2                                                  https://www.ecdc.europa.eu/en/covid-19/data
#> 3                                https://github.com/owid/covid-19-data/tree/master/public/data
#> 4                                                                   https://data.worldbank.org
#> 5                                    https://www.acaps.org/covid19-government-measures-dataset
#> 6  https://www.bsg.ox.ac.uk/research/research-projects/coronavirus-government-response-tracker
#> 7                                                       https://www.apple.com/covid19/mobility
#> 8                                                     https://www.google.com/covid19/mobility/
#> 9                                                                   https://trends.google.com/
#> 10                                               https://github.com/joachim-gassen/tidycovid19
#>     last_data
#> 1  2023-03-09
#> 2  2023-10-23
#> 3  2023-10-26
#> 4  2023-10-27
#> 5  2021-01-04
#> 6  2022-12-31
#> 7  2022-04-12
#> 8  2022-10-15
#> 9  2023-10-22
#> 10 2023-10-27