tidycovid19_variable_definitions {tidycovid19}	R Documentation

Variable Definitions of tidycovid19 Merged Data Frame

Description

List variable definitions and data sources for the data frame created by download_merged_data(). Merge with data(tidycovid19_data_sources) to get full information for each variable, including description of data source and link to URL.

Usage

data(tidycovid19_variable_definitions)

Format

An object of class "data.frame".

Examples

vd <- tidycovid19_variable_definitions
ds <- tidycovid19_data_sources
data_info <- dplyr::left_join(vd, ds, by = c(var_source = "id"))
print(data_info)

[Package tidycovid19 version 0.0.0.9000 Index]