Variable Definitions of tidycovid19 Merged Data Frame

List variable definitions and data sources for the data frame created by download_merged_data(). Merge with data(tidycovid19_data_sources) to get full information for each variable, including description of data source and link to URL.

data(tidycovid19_variable_definitions)

Format

An object of class "data.frame".

Examples

vd <- tidycovid19_variable_definitions
ds <- tidycovid19_data_sources
data_info <- dplyr::left_join(vd, ds, by = c(var_source = "id"))
print(data_info)

#>                  var_name    var_source
#> 1                 country          <NA>
#> 2                   iso3c          <NA>
#> 3                    date          <NA>
#> 4               confirmed      jhu_ccse
#> 5                  deaths      jhu_ccse
#> 6               recovered      jhu_ccse
#> 7                soc_dist     acaps_npi
#> 8                mov_rest     acaps_npi
#> 9              pub_health     acaps_npi
#> 10           gov_soc_econ     acaps_npi
#> 11               lockdown     acaps_npi
#> 12      apple_mtr_driving     apple_mtr
#> 13      apple_mtr_walking     apple_mtr
#> 14      apple_mtr_transit     apple_mtr
#> 15 gcmr_retail_recreation    google_cmr
#> 16  gcmr_grocery_pharmacy    google_cmr
#> 17             gcmr_parks    google_cmr
#> 18  gcmr_transit_stations    google_cmr
#> 19        gcmr_workplaces    google_cmr
#> 20       gcmr_residential    google_cmr
#> 21          gtrends_score google_trends
#> 22  gtrends_country_score google_trends
#> 23                 region         wbank
#> 24                 income         wbank
#> 25             population         wbank
#> 26          land_area_skm         wbank
#> 27            pop_density         wbank
#> 28       pop_largest_city         wbank
#> 29        life_expectancy         wbank
#> 30             gdp_capita         wbank
#> 31              timestamp          <NA>
#>                                                                                                                                                                                                      var_def
#> 1                                                                                                                                                                                               Country name
#> 2                                                                                                                                                        ISO3c country code as defined by ISO 3166-1 alpha-3
#> 3                                                                                                                                                                                              Calendar date
#> 4                                                                                                                                                           Confirmed Covid-19 cases as reported by JHU CSSE
#> 5                                                                                                                                                            Covid-19-related deaths as reported by JHU CSSE
#> 6                                                                                                                                                                Covid-19 recoveries as reported by JHU CSSE
#> 7                                                                                                              Number of social distancing measures reported up to date by ACAPS, net of lifted restrictions
#> 8                                                                                                                   Number of movement restrictions reported up to date by ACAPS, net of lifted restrictions
#> 9                                                                                                                  Number of public health measures reported up to date by ACAPS, net of lifted restrictions
#> 10                                                                                                           Number of social and economic measures reported up to date by ACAPS, net of lifted restrictions
#> 11                                                                                                                      Number of lockdown measures reported up to date by ACAPS, net of lifted restrictions
#> 12                                                                                                       Apple Maps usage for driving directions, as percentage*100 relative to the baseline of Jan 13, 2020
#> 13                                                                                                       Apple Maps usage for walking directions, as percentage*100 relative to the baseline of Jan 13, 2020
#> 14                                                                                                Apple Maps usage for public transit directions, as percentage*100 relative to the baseline of Jan 13, 2020
#> 15              Google Community Mobility Reports data for the frequency that people visit retail and recreation places expressed as a percentage change relative to the baseline period Jan 3 - Feb 6, 2020
#> 16             Google Community Mobility Reports data for the frequency that people visit grocery stores and pharmacies expressed as a percentage change relative to the baseline period Jan 3 - Feb 6, 2020
#> 17                                     Google Community Mobility Reports data for the frequency that people visit parks expressed as a percentage change relative to the baseline period Jan 3 - Feb 6, 2020
#> 18                          Google Community Mobility Reports data for the frequency that people visit transit stations expressed as a percentage change relative to the baseline period Jan 3 - Feb 6, 2020
#> 19                                Google Community Mobility Reports data for the frequency that people visit workplaces expressed as a percentage change relative to the baseline period Jan 3 - Feb 6, 2020
#> 20                        Google Community Mobility Reports data for the frequency that people visit residential places expressed as a percentage change relative to the baseline period Jan 3 - Feb 6, 2020
#> 21                                                                                              Google search volume for the term 'coronavirus', relative across time with the country maximum scaled to 100
#> 22 Country-level Google search volume for the term 'coronavirus' over a period starting Jan 1, 2020, relative across countries with the country having the highest search volume scaled to 100 (time-stable)
#> 23                                                                                                                                              Country region as classified by the World Bank (time-stable)
#> 24                                                                                                                                        Country income group as classified by the World Bank (time-stable)
#> 25                                                                                                         Country population as reported by the World Bank (original identifier 'SP.POP.TOTL', time-stable)
#> 26                                                                                  Country land mass in square kilometers as reported by the World Bank (original identifier 'AG.LND.TOTL.K2', time-stable)
#> 27                                                                                                 Country population density as reported by the World Bank (original identifier 'EN.POP.DNST', time-stable)
#> 28                                                                  Population in the largest metropolian area of the country as reported by the World Bank (original identifier 'EN.URB.LCTY', time-stable)
#> 29                                                           Average life expectancy at birth of country citizens in years as reported by the World Bank (original identifier 'SP.DYN.LE00.IN', time-stable)
#> 30                                                        Country gross domestic product per capita, measured in 2010 US-$ as reported by the World Bank (original identifier 'NY.GDP.PCAP.KD', time-stable)
#> 31                                                                                                                                    Date and time where data has been collected from authoritative sources
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               description
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    <NA>
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    <NA>
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    <NA>
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    <NA>
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    <NA>
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    <NA>
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    The #COVID19 Government Measures Dataset is provided by ACAPS. It puts together measures implemented by governments worldwide in response to the Coronavirus pandemic. Data collection includes secondary data review. The data is reported in event structure with an event reflecting a government measure. Measures are characterized as being either imposing/extending measures or lifting them and categorized in five categories with each category being split up in further sub-categories.
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    The #COVID19 Government Measures Dataset is provided by ACAPS. It puts together measures implemented by governments worldwide in response to the Coronavirus pandemic. Data collection includes secondary data review. The data is reported in event structure with an event reflecting a government measure. Measures are characterized as being either imposing/extending measures or lifting them and categorized in five categories with each category being split up in further sub-categories.
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    The #COVID19 Government Measures Dataset is provided by ACAPS. It puts together measures implemented by governments worldwide in response to the Coronavirus pandemic. Data collection includes secondary data review. The data is reported in event structure with an event reflecting a government measure. Measures are characterized as being either imposing/extending measures or lifting them and categorized in five categories with each category being split up in further sub-categories.
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   The #COVID19 Government Measures Dataset is provided by ACAPS. It puts together measures implemented by governments worldwide in response to the Coronavirus pandemic. Data collection includes secondary data review. The data is reported in event structure with an event reflecting a government measure. Measures are characterized as being either imposing/extending measures or lifting them and categorized in five categories with each category being split up in further sub-categories.
#> 11                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   The #COVID19 Government Measures Dataset is provided by ACAPS. It puts together measures implemented by governments worldwide in response to the Coronavirus pandemic. Data collection includes secondary data review. The data is reported in event structure with an event reflecting a government measure. Measures are characterized as being either imposing/extending measures or lifting them and categorized in five categories with each category being split up in further sub-categories.
#> 12                                                                 Apple's Mobility Trend Reports reflect requests for directions in Apple Maps. The data frame is organized by country-day and its data are expressed as percentages relative to a baseline volume on January 13th, 2020. The data comes in three data frames that you can select by the 'type' parameter. The 'country' data frame contains country-day level data. The 'country_region' data frame provides regional data for regions for which Apple reports regional data. The 'country_city' data frame reports city-level data for cities for which Apple reports this data. Please Note: The original data does not link regions and cities back to countries. This link has been introduced based on geocoding and manual inspection of the results. You can find the according code in the github repository of the {tidycovid19} package but user discretion is advised when using the sub-country level data.
#> 13                                                                 Apple's Mobility Trend Reports reflect requests for directions in Apple Maps. The data frame is organized by country-day and its data are expressed as percentages relative to a baseline volume on January 13th, 2020. The data comes in three data frames that you can select by the 'type' parameter. The 'country' data frame contains country-day level data. The 'country_region' data frame provides regional data for regions for which Apple reports regional data. The 'country_city' data frame reports city-level data for cities for which Apple reports this data. Please Note: The original data does not link regions and cities back to countries. This link has been introduced based on geocoding and manual inspection of the results. You can find the according code in the github repository of the {tidycovid19} package but user discretion is advised when using the sub-country level data.
#> 14                                                                 Apple's Mobility Trend Reports reflect requests for directions in Apple Maps. The data frame is organized by country-day and its data are expressed as percentages relative to a baseline volume on January 13th, 2020. The data comes in three data frames that you can select by the 'type' parameter. The 'country' data frame contains country-day level data. The 'country_region' data frame provides regional data for regions for which Apple reports regional data. The 'country_city' data frame reports city-level data for cities for which Apple reports this data. Please Note: The original data does not link regions and cities back to countries. This link has been introduced based on geocoding and manual inspection of the results. You can find the according code in the github repository of the {tidycovid19} package but user discretion is advised when using the sub-country level data.
#> 15                                                                                                                                                                                               Google's Community Mobility Reports chart movement trends over time across different categories of places such as retail and recreation, groceries and pharmacies, parks, transit stations, workplaces, and residential. They show how visits and length of stays at different places change in percentages compared to a baseline (the median value, for the corresponding day of the week, during the 5-week period Jan 3 to Feb 6, 2020). The data comes in three data frames that you can select by the 'type' parameter. The 'country' data frame contains country-day level data. The 'country_region' data frame provides regional data for the countries for which Google reports regional data. The 'us_county' data frame reports daily data for the U.S. at the county level.
#> 16                                                                                                                                                                                               Google's Community Mobility Reports chart movement trends over time across different categories of places such as retail and recreation, groceries and pharmacies, parks, transit stations, workplaces, and residential. They show how visits and length of stays at different places change in percentages compared to a baseline (the median value, for the corresponding day of the week, during the 5-week period Jan 3 to Feb 6, 2020). The data comes in three data frames that you can select by the 'type' parameter. The 'country' data frame contains country-day level data. The 'country_region' data frame provides regional data for the countries for which Google reports regional data. The 'us_county' data frame reports daily data for the U.S. at the county level.
#> 17                                                                                                                                                                                               Google's Community Mobility Reports chart movement trends over time across different categories of places such as retail and recreation, groceries and pharmacies, parks, transit stations, workplaces, and residential. They show how visits and length of stays at different places change in percentages compared to a baseline (the median value, for the corresponding day of the week, during the 5-week period Jan 3 to Feb 6, 2020). The data comes in three data frames that you can select by the 'type' parameter. The 'country' data frame contains country-day level data. The 'country_region' data frame provides regional data for the countries for which Google reports regional data. The 'us_county' data frame reports daily data for the U.S. at the county level.
#> 18                                                                                                                                                                                               Google's Community Mobility Reports chart movement trends over time across different categories of places such as retail and recreation, groceries and pharmacies, parks, transit stations, workplaces, and residential. They show how visits and length of stays at different places change in percentages compared to a baseline (the median value, for the corresponding day of the week, during the 5-week period Jan 3 to Feb 6, 2020). The data comes in three data frames that you can select by the 'type' parameter. The 'country' data frame contains country-day level data. The 'country_region' data frame provides regional data for the countries for which Google reports regional data. The 'us_county' data frame reports daily data for the U.S. at the county level.
#> 19                                                                                                                                                                                               Google's Community Mobility Reports chart movement trends over time across different categories of places such as retail and recreation, groceries and pharmacies, parks, transit stations, workplaces, and residential. They show how visits and length of stays at different places change in percentages compared to a baseline (the median value, for the corresponding day of the week, during the 5-week period Jan 3 to Feb 6, 2020). The data comes in three data frames that you can select by the 'type' parameter. The 'country' data frame contains country-day level data. The 'country_region' data frame provides regional data for the countries for which Google reports regional data. The 'us_county' data frame reports daily data for the U.S. at the county level.
#> 20                                                                                                                                                                                               Google's Community Mobility Reports chart movement trends over time across different categories of places such as retail and recreation, groceries and pharmacies, parks, transit stations, workplaces, and residential. They show how visits and length of stays at different places change in percentages compared to a baseline (the median value, for the corresponding day of the week, during the 5-week period Jan 3 to Feb 6, 2020). The data comes in three data frames that you can select by the 'type' parameter. The 'country' data frame contains country-day level data. The 'country_region' data frame provides regional data for the countries for which Google reports regional data. The 'us_county' data frame reports daily data for the U.S. at the county level.
#> 21 Data are Google Search Volume (GSV) measures as provided by Google Trends API, with the default search term 'coronavirus'. The data comes in four data frames that you can select by the 'type' parameter and the sample period comprises Jan 1, 2020 up to date. The 'country' data frame lists GSV by country, to assess which country on average uses the search term most often over the sample period. The 'country-day' data frame reports daily search volume data for all countries that show up in the 'country' data frame. Each value is relative within country, meaning that values across countries cannot be compared directly. The 'region' and 'city' data frames list the relative GSV across regions and city within countries when provided by Google Trends. Keep in mind that within each data frame GSV are relative measures with a maximum of 100 indicating the highest search volume. This implies that GSV measures are not comparable across data frames.
#> 22 Data are Google Search Volume (GSV) measures as provided by Google Trends API, with the default search term 'coronavirus'. The data comes in four data frames that you can select by the 'type' parameter and the sample period comprises Jan 1, 2020 up to date. The 'country' data frame lists GSV by country, to assess which country on average uses the search term most often over the sample period. The 'country-day' data frame reports daily search volume data for all countries that show up in the 'country' data frame. Each value is relative within country, meaning that values across countries cannot be compared directly. The 'region' and 'city' data frames list the relative GSV across regions and city within countries when provided by Google Trends. Keep in mind that within each data frame GSV are relative measures with a maximum of 100 indicating the highest search volume. This implies that GSV measures are not comparable across data frames.
#> 23                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The data frame reports current country-level statistics from the World Bank. The regional and income level classifications are also provided by the World Bank. 'life_expectancy' is measured in years at birth and 'gdp_capita' is measured in 2010 US-$. The original World Bank data items are (in the order how they are represented in the data frame) 'SP.POP.TOTL', 'AG.LND.TOTL.K2', 'EN.POP.DNST', 'EN.URB.LCTY', 'SP.DYN.LE00.IN', 'NY.GDP.PCAP.KD'.
#> 24                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The data frame reports current country-level statistics from the World Bank. The regional and income level classifications are also provided by the World Bank. 'life_expectancy' is measured in years at birth and 'gdp_capita' is measured in 2010 US-$. The original World Bank data items are (in the order how they are represented in the data frame) 'SP.POP.TOTL', 'AG.LND.TOTL.K2', 'EN.POP.DNST', 'EN.URB.LCTY', 'SP.DYN.LE00.IN', 'NY.GDP.PCAP.KD'.
#> 25                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The data frame reports current country-level statistics from the World Bank. The regional and income level classifications are also provided by the World Bank. 'life_expectancy' is measured in years at birth and 'gdp_capita' is measured in 2010 US-$. The original World Bank data items are (in the order how they are represented in the data frame) 'SP.POP.TOTL', 'AG.LND.TOTL.K2', 'EN.POP.DNST', 'EN.URB.LCTY', 'SP.DYN.LE00.IN', 'NY.GDP.PCAP.KD'.
#> 26                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The data frame reports current country-level statistics from the World Bank. The regional and income level classifications are also provided by the World Bank. 'life_expectancy' is measured in years at birth and 'gdp_capita' is measured in 2010 US-$. The original World Bank data items are (in the order how they are represented in the data frame) 'SP.POP.TOTL', 'AG.LND.TOTL.K2', 'EN.POP.DNST', 'EN.URB.LCTY', 'SP.DYN.LE00.IN', 'NY.GDP.PCAP.KD'.
#> 27                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The data frame reports current country-level statistics from the World Bank. The regional and income level classifications are also provided by the World Bank. 'life_expectancy' is measured in years at birth and 'gdp_capita' is measured in 2010 US-$. The original World Bank data items are (in the order how they are represented in the data frame) 'SP.POP.TOTL', 'AG.LND.TOTL.K2', 'EN.POP.DNST', 'EN.URB.LCTY', 'SP.DYN.LE00.IN', 'NY.GDP.PCAP.KD'.
#> 28                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The data frame reports current country-level statistics from the World Bank. The regional and income level classifications are also provided by the World Bank. 'life_expectancy' is measured in years at birth and 'gdp_capita' is measured in 2010 US-$. The original World Bank data items are (in the order how they are represented in the data frame) 'SP.POP.TOTL', 'AG.LND.TOTL.K2', 'EN.POP.DNST', 'EN.URB.LCTY', 'SP.DYN.LE00.IN', 'NY.GDP.PCAP.KD'.
#> 29                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The data frame reports current country-level statistics from the World Bank. The regional and income level classifications are also provided by the World Bank. 'life_expectancy' is measured in years at birth and 'gdp_capita' is measured in 2010 US-$. The original World Bank data items are (in the order how they are represented in the data frame) 'SP.POP.TOTL', 'AG.LND.TOTL.K2', 'EN.POP.DNST', 'EN.URB.LCTY', 'SP.DYN.LE00.IN', 'NY.GDP.PCAP.KD'.
#> 30                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The data frame reports current country-level statistics from the World Bank. The regional and income level classifications are also provided by the World Bank. 'life_expectancy' is measured in years at birth and 'gdp_capita' is measured in 2010 US-$. The original World Bank data items are (in the order how they are represented in the data frame) 'SP.POP.TOTL', 'AG.LND.TOTL.K2', 'EN.POP.DNST', 'EN.URB.LCTY', 'SP.DYN.LE00.IN', 'NY.GDP.PCAP.KD'.
#> 31                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   <NA>
#>                                                          url
#> 1                                                       <NA>
#> 2                                                       <NA>
#> 3                                                       <NA>
#> 4                                                       <NA>
#> 5                                                       <NA>
#> 6                                                       <NA>
#> 7  https://www.acaps.org/covid19-government-measures-dataset
#> 8  https://www.acaps.org/covid19-government-measures-dataset
#> 9  https://www.acaps.org/covid19-government-measures-dataset
#> 10 https://www.acaps.org/covid19-government-measures-dataset
#> 11 https://www.acaps.org/covid19-government-measures-dataset
#> 12                    https://www.apple.com/covid19/mobility
#> 13                    https://www.apple.com/covid19/mobility
#> 14                    https://www.apple.com/covid19/mobility
#> 15                  https://www.google.com/covid19/mobility/
#> 16                  https://www.google.com/covid19/mobility/
#> 17                  https://www.google.com/covid19/mobility/
#> 18                  https://www.google.com/covid19/mobility/
#> 19                  https://www.google.com/covid19/mobility/
#> 20                  https://www.google.com/covid19/mobility/
#> 21                                https://trends.google.com/
#> 22                                https://trends.google.com/
#> 23                                https://data.worldbank.org
#> 24                                https://data.worldbank.org
#> 25                                https://data.worldbank.org
#> 26                                https://data.worldbank.org
#> 27                                https://data.worldbank.org
#> 28                                https://data.worldbank.org
#> 29                                https://data.worldbank.org
#> 30                                https://data.worldbank.org
#> 31                                                      <NA>